+++ python/pyautd3/autd.py
@@ -4,7 +4,7 @@ Project: pyautd3
 Created Date: 24/05/2021
 Author: Shun Suzuki
 -----
-Last Modified: 05/01/2023
+Last Modified: 08/01/2023
 Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
 -----
 Copyright (c) 2022 Shun Suzuki. All rights reserved.
@@ -142,10 +142,10 @@ class Geometry:
         self._cnt = cnt
 
     def add_device(self, pos, rot):
-        Base().dll.AUTDAddDevice(self._cnt, pos[0], pos[1], pos[2], rot[0], rot[1], rot[2])
+        return Base().dll.AUTDAddDevice(self._cnt, pos[0], pos[1], pos[2], rot[0], rot[1], rot[2])
 
     def add_device_quaternion(self, pos, q):
-        Base().dll.AUTDAddDeviceQuaternion(self._cnt, pos[0], pos[1], pos[2], q[0], q[1], q[2], q[3])
+        return Base().dll.AUTDAddDeviceQuaternion(self._cnt, pos[0], pos[1], pos[2], q[0], q[1], q[2], q[3])
 
     @ property
     def num_transducers(self):
@@ -238,6 +238,8 @@ class Controller:
         res = []
         handle = c_void_p()
         size = Base().dll.AUTDGetFirmwareInfoListPointer(self.p_cnt, byref(handle))
+        if size < 0:
+            raise Exception('Failed to get firmware version.')
 
         for i in range(size):
             sb = ctypes.create_string_buffer(256)