+++ rust/autd3-driver/src/lib.rs
@@ -1,13 +1,13 @@
 pub mod cpu;
 pub mod defined;
-pub mod driver;
 pub mod error;
 pub mod firmware_version;
 pub mod fpga;
+pub mod operation;
 
 pub use cpu::*;
 pub use defined::*;
-pub use driver::*;
 pub use error::*;
 pub use firmware_version::*;
 pub use fpga::*;
+pub use operation::*;