+++ rust/autd3/src/gain/bessel.rs
@@ -4,31 +4,28 @@
  * Created Date: 02/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 05/01/2023
+ * Last Modified: 09/01/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3_core::{
-    gain::GainProps,
-    geometry::{Geometry, Transducer, UnitQuaternion, Vector3},
-};
+use autd3_core::geometry::{Geometry, Transducer, UnitQuaternion, Vector3};
 
 use autd3_traits::Gain;
 
 /// Gain to produce single focal point
 #[derive(Gain)]
-pub struct Bessel {
-    props: GainProps,
+pub struct Bessel<T: Transducer> {
+    op: T::Gain,
     amp: f64,
     pos: Vector3,
     dir: Vector3,
     theta: f64,
 }
 
-impl Bessel {
+impl<T: Transducer> Bessel<T> {
     /// constructor
     ///
     /// # Arguments
@@ -50,7 +47,7 @@ impl Bessel {
     ///
     pub fn with_duty(pos: Vector3, dir: Vector3, theta: f64, amp: f64) -> Self {
         Self {
-            props: GainProps::new(),
+            op: Default::default(),
             amp,
             pos,
             dir,
@@ -58,7 +55,7 @@ impl Bessel {
         }
     }
 
-    fn calc<T: Transducer>(&mut self, geometry: &Geometry<T>) -> anyhow::Result<()> {
+    fn calc(&mut self, geometry: &Geometry<T>) -> anyhow::Result<()> {
         let dir = self.dir.normalize();
         let v = Vector3::new(dir.y, -dir.x, 0.);
         let theta_v = v.norm().asin();
@@ -75,8 +72,7 @@ impl Bessel {
             let r = rot * r;
             let dist = self.theta.sin() * (r.x * r.x + r.y * r.y).sqrt() - self.theta.cos() * r.z;
             let phase = tr.align_phase_at(dist, sound_speed);
-            self.props.drives[tr.id()].amp = self.amp;
-            self.props.drives[tr.id()].phase = phase;
+            self.op.set_drive(tr.id(), self.amp, phase);
         });
         Ok(())
     }