+++ rust/autd3/src/gain/plane.rs
@@ -4,29 +4,26 @@
  * Created Date: 05/05/2022
  * Author: Shun Suzuki
  * -----
- * Last Modified: 05/01/2023
+ * Last Modified: 09/01/2023
  * Modified By: Shun Suzuki (suzuki@hapis.k.u-tokyo.ac.jp)
  * -----
  * Copyright (c) 2022 Shun Suzuki. All rights reserved.
  *
  */
 
-use autd3_core::{
-    gain::GainProps,
-    geometry::{Geometry, Transducer, Vector3},
-};
+use autd3_core::geometry::{Geometry, Transducer, Vector3};
 
 use autd3_traits::Gain;
 
 /// Gain to produce single focal point
 #[derive(Gain)]
-pub struct Plane {
-    props: GainProps,
+pub struct Plane<T: Transducer> {
+    op: T::Gain,
     amp: f64,
     dir: Vector3,
 }
 
-impl Plane {
+impl<T: Transducer> Plane<T> {
     /// constructor
     ///
     /// # Arguments
@@ -46,19 +43,18 @@ impl Plane {
     ///
     pub fn with_amp(dir: Vector3, amp: f64) -> Self {
         Self {
-            props: GainProps::new(),
+            op: Default::default(),
             amp,
             dir,
         }
     }
 
-    fn calc<T: Transducer>(&mut self, geometry: &Geometry<T>) -> anyhow::Result<()> {
+    fn calc(&mut self, geometry: &Geometry<T>) -> anyhow::Result<()> {
         let sound_speed = geometry.sound_speed;
         geometry.transducers().for_each(|tr| {
             let dist = self.dir.dot(tr.position());
             let phase = tr.align_phase_at(dist, sound_speed);
-            self.props.drives[tr.id()].amp = self.amp;
-            self.props.drives[tr.id()].phase = phase;
+            self.op.set_drive(tr.id(), self.amp, phase);
         });
         Ok(())
     }